% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Distribution}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{sample size}

\item{p}{probability of success}
}
\value{
the proportions of each number of successes,
calculated by dividing the counts in succ.tab by iter.
}
\description{
Binomial Distribution
}
\examples{
mybin(iter=1000,n=10, p=0.7)
}
