% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{Bootstrap Confidence Interval Estimation}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap iterations.}

\item{x}{Vector of data from which to draw samples.}

\item{fun}{The statistic to compute on each bootstrap sample. Default is "mean".}

\item{alpha}{The significance level for the confidence interval. Default is 0.05.}

\item{cx}{Expansion factor for text size in the plot. Default is 1.5.}

\item{...}{Additional arguments to be passed to the hist function for plotting.}
}
\value{
A list containing the bootstrap confidence interval (ci) for the specified statistic,
the chosen statistic (fun), and the original data vector (x).
}
\description{
This function calculates bootstrap confidence intervals for a given statistic
(mean, median, variance, etc.) based on resampling with replacement.
}
\examples{
set.seed(42)
data <- rnorm(100, mean = 10, sd = 2)
boot_result <- myboot(iter = 10000, x = data, fun = "mean", alpha = 0.05)
boot_result$ci

}
