% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{This function generates a Monte Carlo simulation for estimating the distribution of sample means drawn from a uniform distribution.}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{The sample size.}

\item{iter}{The number of iterations for the simulation.}

\item{a}{The lower bound of the uniform distribution (default is 0).}

\item{b}{The upper bound of the uniform distribution (default is 10).}
}
\value{
A histogram and density plot of the sample means along with theoretical density curves.
}
\description{
This function generates a Monte Carlo simulation for estimating the distribution of sample means drawn from a uniform distribution.
}
\examples{
mycltu(100, 1000)
}
