% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{A function representing the likelihood function.}

\item{x}{A vector of data points.}

\item{param}{A vector of parameter values to evaluate the likelihood function at.}

\item{...}{Additional arguments to be passed to the plot function.}
}
\value{
A list containing:
  \item{i}{Index of the maximum likelihood value.}
  \item{parami}{Parameter value at the maximum likelihood.}
  \item{yi}{Maximum likelihood value.}
  \item{slope}{Slope of the likelihood curve at the maximum likelihood point.}
}
\description{
This function is used for estimating the maximum likelihood of parameters
for repeated sampling from the same distribution.
}
\examples{
y <- c(3, 3, 4, 3, 4, 5, 5, 4)
logbin=function(x,param) log(dbinom(x,prob=param,size=20))
mymaxlik(logbin,y,seq(0,1,length=1000),main="Binomial n = 20")

}
