% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot a normal distribution}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean of the normal distribution}

\item{sigma}{standard deviation of the normal distribution}

\item{a}{value to calculate cumulative probability up to}
}
\value{
graph of a normal distribution
}
\description{
Plot a normal distribution
}
\examples{
myncurve(4,2,3)
}
