% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Determine Optimal Number of Tickets to Sell}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The total number of tickets available.}

\item{gamma}{A parameter representing the target revenue.}

\item{p}{The probability of selling a single ticket.}
}
\value{
A named list containing the optimal number of tickets to sell using discrete (`nd`) and continuous (`nc`) distributions, along with the provided parameters.
}
\description{
This function calculates the optimal number of tickets to sell based on the provided parameters using both discrete and continuous distributions.
}
\examples{
# Example usage:
ntickets(400, 0.02, 0.95)
}
