% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwlinear.R
\name{pwlinear}
\alias{pwlinear}
\title{Piecewise linear model}
\usage{
pwlinear(x, xk, coef)
}
\arguments{
\item{x}{x value}

\item{xk}{xk value}

\item{coef}{coefficient}
}
\value{
piecewise linear model
}
\description{
Piecewise linear model
}
\examples{
pwlinear(0, 18, 4)
}
